Add ErrorBoundary for login and signup
