+++ src/components/errorHandling/FallbackComponent.jsx
@@ -1,0 +1,9 @@
+const FallbackComponent = ({ error, resetErrorBoundary }) => (
+    <div role="alert">
+      <h2>Something went wrong.</h2>
+      <p>{error.message}</p>
+      <button onClick={resetErrorBoundary}>Try again</button>
+    </div>
+  )
+
+export default FallbackComponent;
